Fix clippy errors

+++ system-parachains/asset-hubs/asset-hub-kusama/src/lib.rs
@@ -1441,7 +1441,7 @@ mod benches {
 			));
 			let amount = 1_000_000u128;
 			let asset_location =
-				Location::new(0, [PalletInstance(50), GeneralIndex(u32::from(asset_id).into())]);
+				Location::new(0, [PalletInstance(50), GeneralIndex(asset_id.into())]);
 			Asset { id: AssetId(asset_location), fun: Fungible(amount) }
 		}
 	}
@@ -1532,7 +1532,7 @@ mod benches {
 			));
 			let amount = 1_000_000u128;
 			let asset_location =
-				Location::new(0, [PalletInstance(50), GeneralIndex(u32::from(asset_id).into())]);
+				Location::new(0, [PalletInstance(50), GeneralIndex(asset_id.into())]);
 			Asset { id: AssetId(asset_location), fun: Fungible(amount) }
 		}
 	}
